clean up constant array module
