+++ vortex-array/src/arrays/constant/array.rs
@@ -1,0 +1,32 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_scalar::Scalar;
+
+use crate::stats::ArrayStats;
+
+#[derive(Clone, Debug)]
+pub struct ConstantArray {
+    pub(super) scalar: Scalar,
+    pub(super) len: usize,
+    pub(super) stats_set: ArrayStats,
+}
+
+impl ConstantArray {
+    pub fn new<S>(scalar: S, len: usize) -> Self
+    where
+        S: Into<Scalar>,
+    {
+        let scalar = scalar.into();
+        Self {
+            scalar,
+            len,
+            stats_set: Default::default(),
+        }
+    }
+
+    /// Returns the [`Scalar`] value of this constant array.
+    pub fn scalar(&self) -> &Scalar {
+        &self.scalar
+    }
+}
