+++ vortex-array/src/arrays/constant/vtable/array.rs
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_dtype::DType;
+
+use crate::arrays::{ConstantArray, ConstantVTable};
+use crate::stats::StatsSetRef;
+use crate::vtable::ArrayVTable;
+
+impl ArrayVTable<ConstantVTable> for ConstantVTable {
+    fn len(array: &ConstantArray) -> usize {
+        array.len
+    }
+
+    fn dtype(array: &ConstantArray) -> &DType {
+        array.scalar.dtype()
+    }
+
+    fn stats(array: &ConstantArray) -> StatsSetRef<'_> {
+        array.stats_set.to_ref(array.as_ref())
+    }
+}
