+++ vortex-array/src/arrays/constant/vtable/canonical.rs
