+++ vortex-array/src/arrays/constant/vtable/encode.rs
