+++ vortex-array/src/arrays/constant/vtable/pipeline.rs
