+++ vortex-array/src/arrays/constant/vtable/serde.rs
